ENH: Implement ThreadedGenerateData system fully in VideoSource
The ThreadedGenerateData system now works correctly in VideoSource. A subclass
wishing to implement spatial streaming in addition to temporal streaming must
override ThreadedGenerateData to fetch the provided spatial region from the
frames in the requested temporal region of the input and/or output and perform
whatever operation is desired on these frame sub-regions.

One big change that was made while working on VideoSource was the use of frame
numbers rather than offsets when referring to frames from a VideoStream. Under
this idiom, the Head of the underlying frame buffer never moves, but the frame
number (which is treated as an offset internally) just wraps back around as
necessary. This prevents the confusion that comes with worrying about frame
offsets relative to a "current frame" position that could be hard to maintain.