ENH: Make sure output has enough buffers for request at end
If a filter at the end of the pipeline has no requested temporal region set, it
will automatically request it's largest possible (unless the largest possible
is infinite in which case it will request 1). For videos, we want to make sure
that the output video for the last filter has enough frame buffers to hold the
whole requested output, but we don't want this to happen for internal filters.
To solve this, we only adjust the number of internal buffers if the requested
region has not been set yet. This misses the case where the user manually sets
the requested temporal region, but here we assume the user will also manually
set the number of internal buffers (this may not be a perfect assumption).