ENH: Fix propagation of requested spatial regions to output
First, if we just decided to request the largest possible temporal
region by default, we also need to request the largest possible
spatial region for each frame.

Also, we need to make sure that both the cache and frame properly get
the buffered spatial region before allocating. We need to do this in
two steps because the frame may not yet be in the buffered temporal
region (this will happen during the GenerateData phase), so we need
to explicitly set the buffered region on the frame before allocating