ENH: Make all video components const correct
There are two funny places where we need to const_cast. The first is that it is
apparently illegal to call the [] operator on std::map from a const method, so
we need to cast away the constness on the "this" pointer in VideoStream in
order to return the spatial regions from the caches.

The second place is setting input in VideoToVideoFilter. This is caused by a
slight inconsistency in the const correctness of the pipeline design. For the
most part, we assume that a ProcessObject won't change its input, so all
SetInput methods take const pointers. This breaks down, though, when we need to
propagate the requested regions up the pipeline, so we actually need to keep
a non-const pointer to the input under the covers and its then be really
careful to not alter the input in any way other than to change the requested
region.